backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3
rapidfuzz==3.9.0
requests==2.31.0
pytest==8.3.3
python-multipart
